+++ tests/testthat/test-Traceplot.R
@@ -1,14 +1,14 @@
 test_that("traceplot works", {
-  filePath <- system.file("extdata", "epil1.rds", package = "BCClong")
-  fit.BCC <- readRDS(filePath)
+  data("epil1")
+  fit.BCC <- epil1
   temp <- traceplot(fit=fit.BCC, parameter="PPI",
             ylab="pi",xlab="MCMC samples")
-  expect_null(temp)
+  expect_length(temp, 2)
 })
 
 test_that("invailid parameter", {
-  filePath <- system.file("extdata", "epil1.rds", package = "BCClong")
-  fit.BCC <- readRDS(filePath)
+  data("epil1")
+  fit.BCC <- epil1
   expect_error(traceplot(fit=fit.BCC, parameter="POI",
                     ylab="pi",xlab="MCMC samples"))