+++ tests/testthat/test-modelSelection.R
@@ -1,7 +1,7 @@
 test_that("model.selection.criteria works", {
-  filePath <- system.file("extdata", "example1.rds", package = "BCClong")
-  fit.BCC <- readRDS(filePath)
-  res <- model.selection.criteria(fit.BCC, fast_version=1)
+  data("example1")
+  fit.BCC <- example1
+  res <- model.selection.criteria(fit.BCC, fast_version=T)
 
   expect_equal(res$DIC, 3026.2626)
   expect_equal(res$WAIC, 2884.56245)
@@ -9,11 +9,11 @@ test_that("model.selection.criteria works", {
 })
 
 test_that("model.selection.criteria fail", {
-  filePath <- system.file("extdata", "example1.rds", package = "BCClong")
-  fit.BCC <- readRDS(filePath)
+  data("example1")
+  fit.BCC <- example1
 
   expect_error(model.selection.criteria(fit.BCC, fast_version=2),
-               "fast_version should be either 0 or 1")
-  expect_error(model.selection.criteria("a", fast_version=1),
+               "fast_version should be either TRUE or FALSE")
+  expect_error(model.selection.criteria("a", fast_version=T),
                "Object was created without names")
 })